Generic infinite list component.
