+++ packages/app-bird-web/src/components/shared/posts/posts.tsx
@@ -1,7 +1,7 @@
 import {Handler} from "@baqhub/sdk";
 import {DataProvider, RendererOf} from "@baqhub/sdk-react";
-import {Column, tw} from "@baqhub/ui/core/style.js";
-import {PropsWithChildren, useEffect, useRef} from "react";
+import {InfiniteList} from "@baqhub/ui/core/infiniteList.js";
+import {PropsWithChildren} from "react";
 import {LoadingMorePosts} from "./loadingMorePosts.js";
 
 //
@@ -16,12 +16,6 @@ interface PostsProps<T> extends PropsWithChildren {
 }
 
 //
-// Style.
-//
-
-const Layout = tw(Column)``;
-
-//
 // Component.
 //
 
@@ -30,51 +24,14 @@ export function Posts<T>(props: PostsProps<T>) {
   const {isLoadingMore, loadMore} = props;
   const items = getItems();
 
-  //
-  // Load more trigger.
-  //
-
-  const layoutRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const currentLayout = layoutRef.current;
-    if (!currentLayout || !loadMore) {
-      return;
-    }
-
-    const onIntersectionChange: IntersectionObserverCallback = entries => {
-      const intersectionRatio = entries[0]?.intersectionRatio || 0;
-      if (intersectionRatio < 1) {
-        return;
-      }
-
-      console.log("Calling loadMore()");
-      loadMore();
-    };
-
-    const observer = new IntersectionObserver(onIntersectionChange, {
-      rootMargin: "100000px 0px 200px 0px",
-      threshold: 1,
-    });
-
-    observer.observe(currentLayout);
-    return () => {
-      observer.disconnect();
-    };
-  }, [loadMore]);
-
-  //
-  // Render.
-  //
-
   if (items.length === 0) {
     return children;
   }
 
   return (
-    <Layout ref={layoutRef}>
+    <InfiniteList loadMore={loadMore} bottomThreshold={200}>
       {items.map(renderItem)}
       <LoadingMorePosts isLoading={Boolean(isLoadingMore || loadMore)} />
-    </Layout>
+    </InfiniteList>
   );
 }
